rm utils
