+++ nbproject/_schemas.py
@@ -1,9 +1,18 @@
 from datetime import datetime, timezone
 from pathlib import Path
 
-from ._utils import public_fields
 from .dev._initialize import nbproject_id
 from .dev._notebook import Notebook, write_notebook
+
+
+def public_fields(obj):
+    vars_props_dict = {}
+    for key in dir(obj):
+        if key[0] != "_":
+            value = getattr(obj, key)
+            if not callable(value) and value is not None:
+                vars_props_dict[key] = value
+    return vars_props_dict
 
 
 class NBRecord:
